// packages/agent-core/src/types/task-manager.ts
export interface OnBeforeStartContext {
  taskId?: string;
  workspaceId?: string;
}
// ...
export interface OnBeforeStartResult {
  env?: NodeJS.ProcessEnv;
  workspaceInstructions?: string;
}